Scoring + Customer removal
Scoring works. 1 point for giving a customer a drink, +1 if the drink
is the correct drink, +1 if the drink has the correct boba or red bean.

Customers now leave when a order has been fulfilled. Next step is to
properly reset states and have end state work + new sprites to add.